+++ rust/operator-binary/src/catalog/config.rs
@@ -105,6 +105,7 @@ impl CatalogConfig {
     pub async fn from_catalog(
         catalog: &v1alpha1::TrinoCatalog,
         client: &Client,
+        trino_version: u16,
     ) -> Result<CatalogConfig, FromTrinoCatalogError> {
         let catalog_name = catalog
             .meta()
@@ -116,42 +117,42 @@ impl CatalogConfig {
         let mut catalog_config = match &catalog.spec.connector {
             TrinoCatalogConnector::BlackHole(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::DeltaLake(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::GoogleSheet(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::Generic(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::Hive(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::Iceberg(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::Tpcds(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
             TrinoCatalogConnector::Tpch(connector) => {
                 connector
-                    .to_catalog_config(&catalog_name, catalog_namespace, client)
+                    .to_catalog_config(&catalog_name, catalog_namespace, client, trino_version)
                     .await
             }
         }?;