+++ rust/operator-binary/src/catalog/tpch.rs
@@ -13,6 +13,7 @@ impl ToCatalogConfig for TpchConnector {
         catalog_name: &str,
         _catalog_namespace: Option<String>,
         _client: &Client,
+        _trino_version: u16,
     ) -> Result<CatalogConfig, FromTrinoCatalogError> {
         // No additional properties needed
         Ok(CatalogConfig::new(catalog_name.to_string(), CONNECTOR_NAME))